Fix streamlit import recursion for agentic plan
