+++ chat_app_st.py
@@ -10,12 +10,11 @@ import json
 import uuid
 from pathlib import Path
 from gmail_chatbot.agentic_planner import generate_plan
-from gmail_chatbot.agentic_executor import (
-    execute_step,
-    summarize_and_log_agentic_results,  # Added for agentic execution
-    handle_step_limit_reached,
-)
 from gmail_chatbot.email_config import CLAUDE_API_KEY_ENV, load_env
+from gmail_chatbot.agentic_runner import (
+    run_agentic_plan,
+    default_agentic_state_values,
+)
 
 # Module-level logger
 logger = logging.getLogger(__name__)
@@ -46,12 +45,6 @@ st.title("✉️ Gmail Claude Chatbot Assistant")
 
 # --- Agentic Mode Initialization & Controls ---
 # Ensure all agentic mode related keys are initialized first
-default_agentic_state_values = {
-    "current_step_index": 0,
-    "executed_call_count": 0,
-    "accumulated_results": {},
-    "error_messages": []
-}
 if "agentic_mode_enabled" not in st.session_state:
     st.session_state.agentic_mode_enabled = False # This will be the source for the toggle's default
 if "agentic_step_limit" not in st.session_state:
@@ -93,104 +86,6 @@ if st.session_state.agentic_mode_enabled:
 # --- End Agentic Mode Initialization & Controls ---
 
 
-def run_agentic_plan() -> None:
-    """Execute the current agentic plan sequentially with progress feedback."""
-    plan = st.session_state.get("agentic_plan")
-    if not plan:
-        return
-    if not isinstance(plan, list):
-        logger.info("Parsing error: agentic_plan is not a list")
-        return
-
-    agentic_state = st.session_state.get("agentic_state", default_agentic_state_values.copy())
-    step_limit = st.session_state.get("agentic_step_limit", 10)
-
-    progress_bar = st.progress(
-        agentic_state.get("current_step_index", 0) / max(len(plan), 1)
-    )
-
-    while (
-        agentic_state.get("current_step_index", 0) < len(plan)
-        and agentic_state.get("executed_call_count", 0) < step_limit
-    ):
-        idx = agentic_state.get("current_step_index", 0)
-        step_details = plan[idx]
-        if not isinstance(step_details, dict):
-            logger.info("Parsing error: step %s is not a dict", idx)
-            break
-        logger.info(
-            "Starting step %s (%s)",
-            step_details.get("step_id", "N/A"),
-            step_details.get("action_type"),
-        )
-        with st.spinner(f"Executing: {step_details.get('description', 'Working...')}"):
-            try:
-                execution_result = execute_step(step_details, agentic_state)
-            except Exception as e:  # pragma: no cover - defensive
-                st.exception(e)
-                st.error(f"An unexpected error occurred during step execution: {e}")
-                st.session_state.agentic_plan = None
-                st.session_state.agentic_state = default_agentic_state_values.copy()
-                return
-
-        agentic_state = execution_result.get("updated_agentic_state", agentic_state)
-        agentic_state["executed_call_count"] = agentic_state.get("executed_call_count", 0) + 1
-
-        if execution_result.get("status") == "failure":
-            error_msg = (
-                f"Step {idx + 1} ('{step_details.get('step_id', 'Unnamed')}') failed:"
-                f" {execution_result.get('message', 'Unknown error')}"
-            )
-            st.error(error_msg)
-            agentic_state.setdefault("error_messages", []).append(error_msg)
-            summarize_and_log_agentic_results(agentic_state, plan_completed=False)
-            st.session_state.agentic_plan = None
-            st.session_state.agentic_state = default_agentic_state_values.copy()
-            return
-
-        if execution_result.get("requires_user_input", False):
-            st.info(
-                f"Step {idx + 1} requires user input: {execution_result.get('message', '')}"
-            )
-            st.session_state.agentic_state = agentic_state
-            progress_bar.progress((idx + 1) / len(plan))
-            return
-
-        if execution_result.get("status") == "skipped":
-            st.info(execution_result.get("message", "Step skipped"))
-
-        agentic_state["current_step_index"] = idx + 1
-        st.session_state.agentic_state = agentic_state
-        progress_bar.progress(agentic_state["current_step_index"] / len(plan))
-        state_summary = {
-            "executed_call_count": agentic_state.get("executed_call_count"),
-            "current_step_index": agentic_state.get("current_step_index"),
-            "result_keys": list(agentic_state.get("accumulated_results", {}).keys()),
-        }
-        logger.info(
-            "Finished step %s (%s) summary=%s",
-            step_details.get("step_id", "N/A"),
-            step_details.get("action_type"),
-            state_summary,
-        )
-
-    if agentic_state.get("current_step_index", 0) >= len(plan):
-        summarize_and_log_agentic_results(agentic_state, plan_completed=True)
-        st.success("🎉 Agentic plan fully completed!")
-        st.session_state.agentic_plan = None
-        st.session_state.agentic_state = default_agentic_state_values.copy()
-        st.balloons()
-    elif agentic_state.get("executed_call_count", 0) >= step_limit:
-        user_choice = handle_step_limit_reached(agentic_state, step_limit)
-        if user_choice == "continue":
-            st.session_state.agentic_state = agentic_state
-            st.rerun()
-        elif user_choice == "stop":
-            st.session_state.agentic_plan = None
-            st.session_state.agentic_state = default_agentic_state_values.copy()
-            st.rerun()
-        else:
-            st.stop()
 
 # Initialize chat history
